requirements.txt
fastapi
uvicorn
python-dotenv
pandas
requests
